Changes after first iteration of thesis edit

+++ src/app/announcements/announcements.component.ts
@@ -16,11 +16,11 @@ export class AnnouncementsComponent implements OnInit {
 
   constructor(
     private router: Router,
-    private annService: AnnouncementService) {}
+    private annService: AnnouncementService) { }
 
   ngOnInit() {
     this.annService.getAnnouncements()
-      .subscribe( data => {
+      .subscribe(data => {
         this.anns = data;
       });
   }
@@ -30,39 +30,34 @@ export class AnnouncementsComponent implements OnInit {
   }
 
   deleteAnn(announcement: Announcement): void {
-    this.annService.deleteAnn(announcement)
-      .subscribe( data => {
-        this.anns = this.anns.filter(u => u !== announcement);
-      });
+    if (confirm("Are you sure you want to delete?")) {
+      this.annService.deleteAnn(announcement)
+        .subscribe(data => {
+          this.anns = this.anns.filter(u => u !== announcement);
+        });
+    }
   }
-  
-  onOptionsSelected(event){
+
+  onOptionsSelected(event) {
     console.log(event); //option value will be sent as event
-   }
+  }
 
   save(announcement: Announcement) {
-    //date
-    let today = new Date();
-    let dd = String(today.getDate()).padStart(2, '0');
-    let mm = String(today.getMonth() + 1).padStart(2, '0'); //January is 0!
-    let yyyy = today.getFullYear();
-
-    let todayStr = yyyy + '-' + mm + '-' + dd;
-    announcement.dateCreated = todayStr;
-
-    let regexpUrl = new RegExp('^(http:\/\/www\.|https:\/\/www\.|http:\/\/|https:\/\/)?[a-z0-9]+([\-\.]{1}[a-z0-9]+)*\.[a-z]{2,5}(:[0-9]{1,5})?(\/.*)?$');
-    if(!regexpUrl.test(announcement.url) || announcement.anncmntName==""
-    || announcement.url=="")
-    {
-      alert("Ypur input is not valid!")
-    }else{
-    this.annService.editAnn(announcement)
-      .subscribe(res => {
-      alert("Announcement is updated successfully!")
-      }, (err) => {
-        console.log(err);
-        alert(err.error);
-      });
+    if (confirm("Are you sure you want to save?")) {
+      let regexpUrl = new RegExp('^(http:\/\/www\.|https:\/\/www\.|http:\/\/|https:\/\/)?[a-z0-9]+([\-\.]{1}[a-z0-9]+)*\.[a-z]{2,5}(:[0-9]{1,5})?(\/.*)?$');
+      if (!regexpUrl.test(announcement.url) || announcement.anncmntName == ""
+        || announcement.url == "") {
+        alert("Your input is not valid!")
+      } else {
+        this.annService.editAnn(announcement)
+          .subscribe(res => {
+            alert("Announcement is updated successfully!")
+          }, (err) => {
+            console.log(err);
+            alert(err.error);
+          });
+      }
+      this.isEditable = false;
     }
   }
 }
