+++ src/app/myannouncements/myannouncements.component.ts
@@ -30,31 +30,25 @@ anns: Announcement[];
   }
 
   deleteAnn(announcement: Announcement): void {
+    if(confirm("Are you sure you want to delete?")){
     this.annService.deleteAnn(announcement)
       .subscribe( data => {
         this.anns = this.anns.filter(u => u !== announcement);
       });
+    }
   }
   
   onOptionsSelected(event){
     console.log(event); //option value will be sent as event
    }
 
-  save(announcement: Announcement) {
-    //date
-    let today = new Date();
-    let dd = String(today.getDate()).padStart(2, '0');
-    let mm = String(today.getMonth() + 1).padStart(2, '0'); //January is 0!
-    let yyyy = today.getFullYear();
-
-    let todayStr = yyyy + '-' + mm + '-' + dd;
-    announcement.dateCreated = todayStr;
-
+  save(announcement: Announcement) {   
+    if(confirm("Are you sure you want to save?")){
     let regexpUrl = new RegExp('^(http:\/\/www\.|https:\/\/www\.|http:\/\/|https:\/\/)?[a-z0-9]+([\-\.]{1}[a-z0-9]+)*\.[a-z]{2,5}(:[0-9]{1,5})?(\/.*)?$');
     if(!regexpUrl.test(announcement.url) || announcement.anncmntName==""
     || announcement.url=="")
     {
-      alert("Ypur input is not valid!")
+      alert("Your input is not valid!")
     }else{
     this.annService.editAnn(announcement)
       .subscribe(res => {
@@ -64,7 +58,9 @@ anns: Announcement[];
         alert(err.error);
       });
     }
+    this.isEditable=false;
   }
+}
 
    
   public editJob(announcement) {
